+++ ChangeLog.md
@@ -8,6 +8,8 @@ Table:
 
 - Allow empty RowKey in an entity.
 - Fix etag format to be aligned with Azure server.
+- Fix delete none exist table error code and error message, to be aligned with Azure server.
+- Convert entity properties with type "Edm.DateTime" to UTC time, to be aligned with Azure server.
 
 ## 2021.2 Version 3.11.0